alioth: Disable Virtual A/B compression
- since A13, neither XOR compression nor LZ4-compressed VAB snapshots are allowed on A11 VNDK
- userspace snapshots kills incremental updates in recovery which isn't ideal either

```2024-08-04 16:26:02 - ota_from_target_files.py - WARNING : This full OTA is configured to use VABC compression algorithm lz4, which is supported since Android API level 34, but device is launched with 30 . If this full OTA is served to a device running old build, OTA might fail due to unsupported compression parameter. For safety, gz is used because it's supported since day 1.
2024-08-04 16:26:02 - ota_from_target_files.py - ERROR   : VABC XOR not supported on this vendor, disabling```

Change-Id: I5fac49ccdb8ca828399f5ee7dbce73df04221f70
Signed-off-by: Omkar Chandorkar <gotenksIN@aospa.co>